requirements.txt
google-cloud-aiplatform
google-cloud-storage
xgboost
scikit-learn
pandas
numpy